requirements.txt
Django>=5.0.0,<6.0.0
mysqlclient>=2.2.0
Pillow>=10.0.0
djangorestframework>=3.14.0
